Add support for viewsreference 8.x-2.0-alpha4
