+++ composer.json
@@ -1,5 +1,5 @@
 {
-  "name": "drupal/graphql-views",
+  "name": "drupal/graphql_views",
   "type": "drupal-module",
   "description": "Exposes your Drupal Views data model through a GraphQL schema.",
   "homepage": "http://drupal.org/project/graphql_views",
