+++ src/Plugin/Deriver/ViewDeriverBase.php
@@ -10,6 +10,7 @@ use Drupal\Core\Plugin\Discovery\ContainerDeriverInterface;
 use Drupal\graphql_views\Plugin\views\row\GraphQLEntityRow;
 use Drupal\graphql_views\Plugin\views\row\GraphQLFieldRow;
 use Drupal\graphql\Utility\StringHelper;
+use Drupal\graphql_views\ViewDeriverHelperTrait;
 use Drupal\views\Plugin\views\display\DisplayPluginInterface;
 use Drupal\views\ViewEntityInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -18,6 +19,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * Base class for graphql view derivers.
  */
 abstract class ViewDeriverBase extends DeriverBase implements ContainerDeriverInterface {
+  use ViewDeriverHelperTrait {getRowResolveType as private traitGetRowResolveType;}
   /**
    * The entity type manager.
    *
@@ -66,52 +68,6 @@ abstract class ViewDeriverBase extends DeriverBase implements ContainerDeriverIn
   }
 
   /**
-   * Check if a pager is configured.
-   *
-   * @param \Drupal\views\Plugin\views\display\DisplayPluginInterface $display
-   *   The display configuration.
-   *
-   * @return bool
-   *   Flag indicating if the view is configured with a pager.
-   */
-  protected function isPaged(DisplayPluginInterface $display) {
-    $pagerOptions = $display->getOption('pager');
-    return isset($pagerOptions['type']) && in_array($pagerOptions['type'], ['full', 'mini']);
-  }
-
-  /**
-   * Get the configured default limit.
-   *
-   * @param \Drupal\views\Plugin\views\display\DisplayPluginInterface $display
-   *   The display configuration.
-   *
-   * @return int
-   *   The default limit.
-   */
-  protected function getPagerLimit(DisplayPluginInterface $display) {
-    $pagerOptions = $display->getOption('pager');
-    return NestedArray::getValue($pagerOptions, [
-      'options', 'items_per_page',
-    ]) ?: 0;
-  }
-
-  /**
-   * Get the configured default offset.
-   *
-   * @param \Drupal\views\Plugin\views\display\DisplayPluginInterface $display
-   *   The display configuration.
-   *
-   * @return int
-   *   The default offset.
-   */
-  protected function getPagerOffset(DisplayPluginInterface $display) {
-    $pagerOptions = $display->getOption('pager');
-    return NestedArray::getValue($pagerOptions, [
-      'options', 'offset',
-    ]) ?: 0;
-  }
-
-  /**
    * Retrieves the entity type id of an entity by its base or data table.
    *
    * @param string $table
@@ -137,177 +93,20 @@ abstract class ViewDeriverBase extends DeriverBase implements ContainerDeriverIn
     return !empty($this->dataTables[$table]) ? $this->dataTables[$table] : NULL;
   }
 
+
   /**
    * Retrieves the type the view's rows resolve to.
    *
    * @param \Drupal\views\ViewEntityInterface $view
    *   The view entity.
-   * @param $displayId
-   *   The id of the current display.
+   * @param string $displayId
+   *   Interface plugin manager.
    *
    * @return null|string
    *   The name of the type or NULL if the type could not be derived.
    */
   protected function getRowResolveType(ViewEntityInterface $view, $displayId) {
-    /** @var \Drupal\graphql_views\Plugin\views\display\GraphQL $display */
-    $display = $this->getViewDisplay($view, $displayId);
-    $rowPlugin = $display->getPlugin('row');
-
-    if ($rowPlugin instanceof GraphQLFieldRow) {
-      return StringHelper::camelCase($display->getGraphQLRowName());
-    }
-
-    if ($rowPlugin instanceof GraphQLEntityRow) {
-      $executable = $view->getExecutable();
-      $executable->setDisplay($displayId);
-
-      if ($entityType = $executable->getBaseEntityType()) {
-        $typeName = $entityType->id();
-        $typeNameCamel = StringHelper::camelCase($typeName);
-        if ($this->interfaceExists($typeNameCamel)) {
-          $filters = $executable->getDisplay()->getOption('filters');
-          $dataTable = $entityType->getDataTable();
-          $bundleKey = $entityType->getKey('bundle');
-
-          foreach ($filters as $filter) {
-            $isBundleFilter = $filter['table'] == $dataTable && $filter['field'] == $bundleKey;
-            $isSingleValued = is_array($filter['value']) && count($filter['value']) == 1;
-            $isExposed = isset($filter['exposed']) && $filter['exposed'];
-            if ($isBundleFilter && $isSingleValued && !$isExposed) {
-              $bundle = reset($filter['value']);
-              $typeName .= "_$bundle";
-              break;
-            }
-          }
-
-          return StringHelper::camelCase($typeName);
-        }
-      }
-
-      return 'Entity';
-    }
-
-    return NULL;
+    return $this->traitGetRowResolveType($view, $displayId, $this->interfacePluginManager);
   }
 
-  /**
-   * Check if a certain interface exists.
-   *
-   * @param string $interface
-   *   The GraphQL interface name.
-   *
-   * @return bool
-   *   Boolean flag indicating if the interface exists.
-   */
-  protected function interfaceExists($interface) {
-    return (bool) array_filter($this->interfacePluginManager->getDefinitions(), function($definition) use ($interface) {
-      return $definition['name'] === $interface;
-    });
-  }
-
-  /**
-   * Returns a view display object.
-   *
-   * @param \Drupal\views\ViewEntityInterface $view
-   *   The view object.
-   * @param string $displayId
-   *   The display ID to use.
-   *
-   * @return \Drupal\views\Plugin\views\display\DisplayPluginInterface
-   *   The view display object.
-   */
-  protected function getViewDisplay(ViewEntityInterface $view, $displayId) {
-    $viewExecutable = $view->getExecutable();
-    $viewExecutable->setDisplay($displayId);
-    return $viewExecutable->getDisplay();
-  }
-
-  /**
-   * Returns a view style object.
-   *
-   * @param \Drupal\views\ViewEntityInterface $view
-   *   The view object.
-   * @param string $displayId
-   *   The display ID to use.
-   *
-   * @return \Drupal\views\Plugin\views\style\StylePluginBase
-   *   The view style object.
-   */
-  protected function getViewStyle(ViewEntityInterface $view, $displayId) {
-    $viewExecutable = $view->getExecutable();
-    $viewExecutable->setDisplay($displayId);
-    return $viewExecutable->getStyle();
-  }
-
-  /**
-   * Returns cache metadata plugin definitions.
-   *
-   * @param \Drupal\views\ViewEntityInterface $view
-   *   The view object.
-   *
-   * @return array
-   *   The cache metadata definitions for the plugin definition.
-   */
-  protected function getCacheMetadataDefinition(ViewEntityInterface $view, DisplayPluginInterface $display) {
-    $metadata = $display->getCacheMetadata()
-      ->addCacheTags($view->getCacheTags())
-      ->addCacheContexts($view->getCacheContexts())
-      ->mergeCacheMaxAge($view->getCacheMaxAge());
-
-    return [
-      'schema_cache_tags' => $metadata->getCacheTags(),
-      'schema_cache_max_age' => $metadata->getCacheMaxAge(),
-      'response_cache_contexts' => array_filter($metadata->getCacheContexts(), function ($context) {
-        // Don't emit the url cache contexts.
-        return $context !== 'url' && strpos($context, 'url.') !== 0;
-      }),
-    ];
-  }
-
-  /**
-   * Returns information about view arguments (contextual filters).
-   *
-   * @param array $viewArguments
-   *   The "arguments" option of a view display.
-   *
-   * @return array
-   *   Arguments information keyed by the argument ID. Subsequent array keys:
-   *     - index: argument index.
-   *     - entity_type: target entity type.
-   *     - bundles: target bundles (can be empty).
-   */
-  protected function getArgumentsInfo(array $viewArguments) {
-    $argumentsInfo = [];
-
-    $index = 0;
-    foreach ($viewArguments as $argumentId => $argument) {
-      $info = [
-        'index' => $index,
-        'entity_type' => NULL,
-        'bundles' => [],
-      ];
-
-      if (isset($argument['entity_type']) && isset($argument['entity_field'])) {
-        $entityType = $this->entityTypeManager->getDefinition($argument['entity_type']);
-        if ($entityType) {
-          $idField = $entityType->getKey('id');
-          if ($idField === $argument['entity_field']) {
-            $info['entity_type'] = $argument['entity_type'];
-            if (
-              $argument['specify_validation'] &&
-              strpos($argument['validate']['type'], 'entity:') === 0 &&
-              !empty($argument['validate_options']['bundles'])
-            ) {
-              $info['bundles'] = $argument['validate_options']['bundles'];
-            }
-          }
-        }
-      }
-
-      $argumentsInfo[$argumentId] = $info;
-      $index++;
-    }
-
-    return $argumentsInfo;
-  }
 }
